+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/InternalResource.java
@@ -484,7 +484,43 @@ public interface InternalResource {
          *
          * @since 23.1
          */
-        WINDOWS("windows");
+        WINDOWS("windows"),
+
+        /**
+         * Represents an unsupported operating system.
+         * <p>
+         * To enable execution on unsupported platforms, the system property
+         * {@code -Dpolyglot.engine.allowUnsupportedPlatform=true} must be explicitly set. If this
+         * property is not set and the platform is unsupported, the {@link #getCurrent()} method
+         * will throw an {@link IllegalStateException}.
+         *
+         * @since 26.0
+         */
+        UNSUPPORTED("unsupported");
+
+        private static final String PROPERTY_ALLOW_UNSUPPORTED_PLATFORM = "polyglot.engine.allowUnsupportedPlatform";
+
+        private static volatile Boolean allowsUnsupportedPlatformValue;
+
+        private static boolean allowsUnsupportedPlatform() {
+            Boolean res = allowsUnsupportedPlatformValue;
+            if (res == null) {
+                synchronized (OS.class) {
+                    res = allowsUnsupportedPlatformValue;
+                    if (res == null) {
+                        res = Boolean.getBoolean(OS.PROPERTY_ALLOW_UNSUPPORTED_PLATFORM);
+                        if (!ImageInfo.inImageBuildtimeCode()) {
+                            /*
+                             * Avoid caching the property value during image build time, as it would
+                             * require resetting it in the image heap later.
+                             */
+                            allowsUnsupportedPlatformValue = res;
+                        }
+                    }
+                }
+            }
+            return res;
+        }
 
         private final String id;
 
@@ -517,8 +553,12 @@ public interface InternalResource {
                 return DARWIN;
             } else if (os.toLowerCase().startsWith("windows")) {
                 return WINDOWS;
+            } else if (allowsUnsupportedPlatform()) {
+                return UNSUPPORTED;
             } else {
-                throw CompilerDirectives.shouldNotReachHere("Unsupported OS name " + os);
+                throw new IllegalStateException(String.format("Unsupported operating system: '%s'. " +
+                                "If you want to continue using this unsupported platform, set the system property '-D%s=true'. " +
+                                "Note that unsupported platforms require additional command line options to be functional.", os, PROPERTY_ALLOW_UNSUPPORTED_PLATFORM));
             }
         }
     }
@@ -542,7 +582,19 @@ public interface InternalResource {
          *
          * @since 23.1
          */
-        AMD64("amd64");
+        AMD64("amd64"),
+
+        /**
+         * Represents an unsupported architecture.
+         * <p>
+         * To enable execution on unsupported platforms, the system property
+         * {@code -Dpolyglot.engine.allowUnsupportedPlatform=true} must be explicitly set. If this
+         * property is not set and the platform is unsupported, the {@link #getCurrent()} method
+         * will throw an {@link IllegalStateException}.
+         *
+         * @since 26.0
+         */
+        UNSUPPORTED("unsupported");
 
         private final String id;
 
@@ -573,7 +625,15 @@ public interface InternalResource {
             return switch (arch) {
                 case "amd64", "x86_64" -> AMD64;
                 case "aarch64", "arm64" -> AARCH64;
-                default -> throw CompilerDirectives.shouldNotReachHere("Unsupported CPU architecture " + arch);
+                default -> {
+                    if (OS.allowsUnsupportedPlatform()) {
+                        yield UNSUPPORTED;
+                    } else {
+                        throw new IllegalStateException(String.format("Unsupported CPU architecture: '%s'. " +
+                                        "If you want to allow unsupported CPU architectures (which will cause the fallback Truffle runtime to be used and may disable some language features), " +
+                                        "set the system property '-D%s=true'.", arch, OS.PROPERTY_ALLOW_UNSUPPORTED_PLATFORM));
+                    }
+                }
             };
         }
     }